client/requirements.txt
watchdog
gitpython
peewee>=3.11
